#529: the branch could not pass CI, and three outcomes still disagreed with reality
All from the GPT second opinion, which returned DO-NOT-SHIP where Grok returned
SHIP. The verdict driver was the one I should have caught myself.

0. THE BRANCH COULD NOT MERGE. My own fixes pushed internal/providers/cookies.go
   to 814 lines, over the repo's enforced 800-line ceiling. `make repo-hygiene`
   fails, so CI fails. I reported this branch as ready without running the gate
   its own repository enforces.

   Fixed by moving browserCookiesForURLWithOutcome to cookie_outcome.go: 814 ->
   690 and 176 -> 333. That is a real seam, not line-shaving. The function now
   returns the outcome triple, so it belongs beside the type that names those
   outcomes and the reporter that renders them; cookie_outcome.go becomes the one
   place where "what happened" is both decided and said, and cookies.go keeps the
   browser and profile machinery.

1. A LATE OPT-OUT LEFT THE OUTCOME LYING. permittedAfterRead re-asks for consent
   AFTER the seconds-long read, so it can empty the list at the last moment. It
   was applied to the cookies alone, leaving the outcome saying "found" beside no
   cookies -- a caller acting on the outcome would report a successful read that
   returned nothing. That is the ambiguity #529 exists to remove, reintroduced by
   the consent gate rather than the reader. Now reported as declined, because the
   stores may well have held cookies and the user simply withdrew permission.

   Flagged independently by BOTH reviewers. I fixed neither the first time.

2. A PARTIAL FAILURE WAS REPORTED AS "YOU ARE NOT LOGGED IN". kooky reads several
   stores and can return cookies from one while reporting an error from another,
   so the error branch -- which requires len(cookies)==0 -- never fired. If domain
   filtering then left nothing, the answer was outcomeNoMatch: a confident claim
   about the user's account, made from a read that partly failed.

3. A PORT-ONLY URL PASSED VALIDATION. The check read u.Host == "". "https://:443/"
   has a NON-EMPTY Host (":443") and an EMPTY Hostname, so it reached the cookie
   stores with an empty domain suffix -- matching whatever that happens to match
   rather than refusing a URL that names no host. Now checks Hostname().

Sabotage-verified: restoring u.Host lets every port-only URL through to the
stores. Three tests added for the three defects; providers passes; the file-size
gate passes with a true exit 0, read from the script rather than from a pipe.
